1. KPP1-127, 2. Fix a bug in router, it will go into a dead loop when clear database.

+++ KDS/app/src/main/java/com/bematechus/kds/KDSDBStatistic.java
@@ -9,6 +9,7 @@ import com.bematechus.kdslib.ConditionBase;
 import com.bematechus.kdslib.ConditionOneTime;
 import com.bematechus.kdslib.ConditionStatistic;
 import com.bematechus.kdslib.DateSlots;
+import com.bematechus.kdslib.KDSConst;
 import com.bematechus.kdslib.KDSDBBase;
 import com.bematechus.kdslib.KDSDataCondiment;
 import com.bematechus.kdslib.KDSDataCondiments;
@@ -150,14 +151,14 @@ public class KDSDBStatistic extends KDSDBBase {
 
             this.executeDML(sql);
             //set the finished time
-            java.util.Date dt = new java.util.Date();
-            sql = "update orders set finishedtime='" + KDSUtil.convertDateToString(dt) + "' where guid='" + guid + "'";//+ KDSUtil.ConvertIntToString(nOrderID);
-            this.executeDML(sql);
+            if (KDSConst.ENABLE_FEATURE_STATISTIC) {
+                java.util.Date dt = new java.util.Date();
+                sql = "update orders set finishedtime='" + KDSUtil.convertDateToString(dt) + "' where guid='" + guid + "'";//+ KDSUtil.ConvertIntToString(nOrderID);
+                this.executeDML(sql);
 
-            KDSDataItems items = order.getItems();
-
-            itemsAdd(items);
-
+                KDSDataItems items = order.getItems();
+                itemsAdd(items);
+            }
             //this.finishTransaction(bTransactionByMe);
             this.commitTransaction(bTransactionByMe);
             return true;
@@ -176,6 +177,8 @@ public class KDSDBStatistic extends KDSDBBase {
     }
     private boolean itemsAdd(KDSDataItems items)
     {
+        if (!KDSConst.ENABLE_FEATURE_STATISTIC)
+            return true;
         int ncount = items.getCount();
         String sql = "";
         try
@@ -200,6 +203,8 @@ public class KDSDBStatistic extends KDSDBBase {
     }
     private boolean condimentsAdd(KDSDataCondiments condiments)
     {
+        if (!KDSConst.ENABLE_FEATURE_STATISTIC)
+            return true;
         int ncount = condiments.getCount();
         String sql = "";
         try
@@ -237,17 +242,18 @@ public class KDSDBStatistic extends KDSDBBase {
 //            String itemguid = getString(c,0);
 //            itemDelete(itemguid);
 //        }
-        sql = String.format("delete from condiments where condiments.itemguid in (select guid from items where items.orderguid='%s')", guid);
-        if (!this.executeDML(sql))
-            return false;
-        sql = String.format("delete from messages where ObjType=1 and messages.ObjGUID in (select guid from items where items.orderguid='%s')", guid);
-        if (!this.executeDML(sql))
-            return false;
+        if (KDSConst.ENABLE_FEATURE_STATISTIC) {
+            sql = String.format("delete from condiments where condiments.itemguid in (select guid from items where items.orderguid='%s')", guid);
+            if (!this.executeDML(sql))
+                return false;
+            sql = String.format("delete from messages where ObjType=1 and messages.ObjGUID in (select guid from items where items.orderguid='%s')", guid);
+            if (!this.executeDML(sql))
+                return false;
 
-        sql = String.format("delete from items where items.orderguid='%s'", guid);
-        if (!this.executeDML(sql))
-            return false;
-
+            sql = String.format("delete from items where items.orderguid='%s'", guid);
+            if (!this.executeDML(sql))
+                return false;
+        }
         return true;
     }
     public boolean itemDelete(String itemguid)//int nID)
@@ -1115,6 +1121,8 @@ public class KDSDBStatistic extends KDSDBBase {
     /************************************************************************/
     TimeSlotOrderReport createItems_Report(ConditionStatistic condition)
     {
+        if (!KDSConst.ENABLE_FEATURE_STATISTIC)
+            return null;
 
         String sql = "";
 
@@ -1196,8 +1204,11 @@ public class KDSDBStatistic extends KDSDBBase {
         Date dt = c.getTime();
         String strDate = KDSUtil.convertDateToDbString(dt);
         boolean bstartedbyme = this.startTransaction();
-        String sql = String.format("delete from items where items.orderguid in (select guid from orders where date(finishedtime)<='%s')", strDate);
-        this.executeDML(sql);
+        String sql = "";
+        if (KDSConst.ENABLE_FEATURE_STATISTIC) {
+            sql = String.format("delete from items where items.orderguid in (select guid from orders where date(finishedtime)<='%s')", strDate);
+            this.executeDML(sql);
+        }
         sql = String.format("delete from orders where date(finishedtime)<='%s'", strDate);
         this.executeDML(sql);
         this.finishTransaction(bstartedbyme);
@@ -1227,8 +1238,10 @@ public class KDSDBStatistic extends KDSDBBase {
     {
         String sql = String.format("delete from orders where guid='%s'", orderGuid);
         this.executeDML(sql);
-        sql = String.format("delete from items where orderguid='%s'", orderGuid);
-        this.executeDML(sql);
+        if (KDSConst.ENABLE_FEATURE_STATISTIC) {
+            sql = String.format("delete from items where orderguid='%s'", orderGuid);
+            this.executeDML(sql);
+        }
         return true;
     }
 
@@ -1437,13 +1450,14 @@ public class KDSDBStatistic extends KDSDBBase {
         boolean bStartedByMe = this.startTransaction();
         String sql = "delete from orders";
         this.executeDML(sql);
-        sql = "delete from items";
-        this.executeDML(sql);
-        sql = "delete from condiments";
-        this.executeDML(sql);
-        sql = "delete from messages";
-        this.executeDML(sql);
-
+        if (KDSConst.ENABLE_FEATURE_STATISTIC) {
+            sql = "delete from items";
+            this.executeDML(sql);
+            sql = "delete from condiments";
+            this.executeDML(sql);
+            sql = "delete from messages";
+            this.executeDML(sql);
+        }
         this.finishTransaction(bStartedByMe);
